clipboard: refactored interface.
